requirements.txt
matplotlib==3.7.1
numpy==1.21
opencv-python==4.7.0.72
openslide-python==1.2.0
pandas==2.0.2
Pillow==9.4.0
PyYAML==6.0
scipy==1.10.1
seaborn==0.12.2
sympy==1.11.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
ultralytics==8.0.118
urllib3==2.0.3
python-dateutil>=2.8.2
requests>=2.23.0
numexpr>=2.7.3
typing-extensions>=4.7.1